clauncher/requirements.txt
paho-mqtt>=1.1
gdown==3.10.1